\name{TabulaMurisData}
\docType{package}
\alias{TabulaMurisData}
\title{
    TabulaMurisData
}
\description{
    This ExperimentHub package provides access to the 10x (droplet) and
    SmartSeq2 (on FACS-sorted cells) single-cell RNA-seq data from the Tabula
    Muris Consortium (http://tabula-muris.ds.czbiohub.org/)
}
\author{
    Charlotte Soneson
}
\references{
    The Tabula Muris Consortium:
    Single-cell transcriptomics of 20 mouse organs creates a \emph{Tabula Muris}.
    Nature doi:10.1038/s41586-018-0590-4 (2018).
}
