\name{TabulaMurisData}
\docType{package}
\alias{TabulaMurisData}
\title{
    TabulaMurisData
}
\description{
    Single-cell RNA-seq data from the Tabula Muris Consortium.
}

\details{
    This ExperimentHub package provides access to the 10x (droplet) and
    SmartSeq2 (on FACS-sorted cells) single-cell RNA-seq data from the Tabula
    Muris Consortium (http://tabula-muris.ds.czbiohub.org/)
}
\author{
    Charlotte Soneson
}
\references{
    The Tabula Muris Consortium, Stephen R. Quake, Tony Wyss-Coray,
    Spyros Darmanis:
    Single-cell transcriptomic characterization of 20 organs and tissues from
    individual mice creates a \emph{Tabula Muris}.
    bioRxiv doi:10.1101/237446 (2017).
}
